Using Section component for About section. De-coupled code to make re-usable components
