+++ src/css/util/section.css
@@ -1,10 +1,9 @@
 .section-parent {
+	margin: auto;
 	margin-bottom: 2.25rem;
 	border-radius: 1.75rem;
 	padding: 0rem;
 	background: white;
-
-	max-width: 100%;
 }
 
 .section-header-container {
